requirements.txt
matplotlib
numpy
pandas
Pillow
pytorch-lightning
scikit-learn
segmentation-models-pytorch
tensorboard 
torch
torchvision